Added food.nourishment
This is an easy alternative to food.saturation. It uses a registry that indexes the values present at https://minecraft.gamepedia.com/Food#Nourishment_value.
As a result of adding this, food.saturation logic has changed. It defaults to -1.0F, which ignores the value and uses food.nourishment instead. Setting food.saturation manually overrides food.nourishment.